Moved TempMod routine into subpackage
it will get more classes soon
