+++ application/src/main/java/org/togetherjava/tjbot/commands/moderation/temp/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package offers classes dealing with temporary moderation actions, such as temporary bans.
+ */
+package org.togetherjava.tjbot.commands.moderation.temp;
